announcement,announcementToProduct crud functionality and settings added

+++ ProductsDistribution.Core/AnnouncementToProduct/Models/AnnouncementToProductDTO.cs
@@ -9,7 +9,7 @@ namespace ProductsDistribution.Core.AnnouncementToProduct.Models
     {
         public int announcement_to_product_id { get; set; }
 
-        public float max_quantity { get; set; }
+        public double max_quantity { get; set; }
 
         public int announcement_id { get; set; }
